pyproject.toml
[tool.poetry]
name = "ibkr"
version = "0.1.0"
description = "APIs for interacting with IBKR account and getting market data."
authors = ["Ivan Kyulev <ivandkyulev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
